% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{combine_hwdata}
\alias{combine_hwdata}
\title{Combine data}
\usage{
combine_hwdata(flip = TRUE)
}
\description{
Combine all counts into one dataframe.
}
\examples{
\dontrun{
 write.table(combine_hwdata(),
   file=file.path("inst", "extdata", "combined_YRICEU.txt"),
   quote=FALSE, row.names=FALSE, sep="\\t")
}
}

