% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{eve102_data}
\alias{eve102_data}
\title{Retrieve a list of all datasets, or the path of a specific dataset}
\usage{
eve102_data(id = NULL)
}
\arguments{
\item{id}{the \code{id} to specify which dataset to return the filepath to.
With no argument supplied, this returns a dataframe of all datasets.}
}
\description{
When run without arguments, this function returns a list of all datasets in
the \code{eve102} package, with the description, filename, and filepath of
each dataset. To retrieve the filepath of a specific dataset (e.g. to use
this path to load it), call \code{eve102_data()} with the \code{id} of the
dataset you wish to have the path to.
}
\examples{
eve102_data()  # retrieve all datasets
eve102_data(2)  # retrieve the path to the second dataset
}

