% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{eve102_status}
\alias{eve102_status}
\title{Check GitHub to see if the eve102 package is out of date.}
\usage{
eve102_status()
}
\description{
This function checks whether the eve102 package is out of date. You should
run this periodically throughout the quarter. We will tell you via email
when to update the package; this is just a way to ensure you're using the
most recent version.
}
\details{
DO NOT PUT THIS FUNCTION IN YOUR RMARKDOWN FILES.  Calling this function
each time you knit your RMarkdown files will make many unnecessary requests
to GitHub. Check by entering \code{eve102_status()} directly in your terminal
only.
}

