% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{tidy_hwdata}
\alias{tidy_hwdata}
\title{Tidy up raw genptype data from PLINK (internal function)}
\usage{
tidy_hwdata(file, flip = TRUE)
}
\description{
For extracting the genotype counts from the original raw data from PLINK.
}
\examples{
\dontrun{
 # this is how tidy versions were created (run in package dir)
 files <- sapply(RAWFILES, eve102_data, include_raw=TRUE)
 for (i in seq_along(files)) {
   outfile <- file.path("inst", "extdata", paste0(names(files)[i], ".txt.gz"))
   outcon <- gzfile(outfile, "w")
   write.table(tidy_hwdata(file.path("inst", "extdata", basename(files[i]))), 
               file=outcon, sep="\\t", quote=FALSE, row.names=FALSE)
   close(outcon)
 }
}
}
