% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{tutorials}
\alias{tutorials}
\title{Browse EVE102 R tutorials, or open specific tutorial}
\usage{
tutorials(num = NULL)
}
\arguments{
\item{num}{tutorial number (numbers 1-3).}
}
\description{
Running \code{tutorials()} without arguments opens a list of all 
available EVE102 tutorials. Entering a specific \code{num} opens that
tutorial.
}
